qcacld-3.0: Fix csr_get_current_band() context param
Currently csr_get_current_band() takes a tHalHandle context param.
However csr is internal to the UMAC, and hence it should be using the
"real" context pointer type instead of the opaque handle, so update
the API to expect a struct mac_context pointer..

This is part of an ongoing campaign to remove or replace all instances
of tHalHandle.

Change-Id: I93a3950b9ee26284ddfdf167b900d2f6983625f7
CRs-Fixed: 2347655
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>