qcacld-3.0: Fix serialization command list full issue
In __wlan_hdd_cfg80211_del_station while disconnecting all the
connected stations, driver send both disassoc and deauth and wait
for disconnect to complete for the STA. Once disconnect is completed
for the 1st STA, it continue to disconnect the next STA.

Now as both disassoc and deauth were queued and both of them set the
wait event resulting in wait event set twice for a disconnect
indication. This result in wait completion for the next STA disconnect
as well. This lead dynamo effect and multiple STA disassoc and deauth
commands are added in the serialization queue filling it up.

Fix this by sending only deauth STA req and removing the duplicate
disassoc STA req. Thus the next command will be queued only when
disconnect is completed for 1st STA.

Change-Id: I4ae6c578a3177a2a6ff75e6e75606e8c5e2f81a0
CRs-Fixed: 2350557
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>