qcacld-3.0: Verify the nss config with rf chains supported
The driver fills the nss chains params in the vdev during
vdev create. The nss and chain params are filled by the user
in the ini file, but it may happen that the user filled the
nss param for that particular vdev as 2, and the number of
supported rf chains are 1, which is sent to the driver by FW,
so that would result in nss as 2, and chains as 1, which cannot
be supported.

Fix is to downgrade the nss to the number of chains supported by
the FW.

Change-Id: Ifa3ee74dbe5b8d19c6dc11ff64e42dbffadaa018
CRs-Fixed: 2354205
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>