qcacld-3.0: Enable power save with user cfg request
Currently power save enable request user command is not
processed if the station is not in connected state and the
configuration is not saved hence station do not enter into
power save mode until the user command is issued after
connection.
Save the power save user configuration setting if the station
is not connected and start the auto power save timer for one
second timeout value after the station is connected to enable
the power save.

Change-Id: I6e17a8e989ca731ca008dcb46ca79c3a181ddff0
CRs-Fixed: 2371045
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>