qcacld-3.0: Revise wma_mcs_rate_match() implementation
Currently some of the parameter names used in the implementation of
wma_mcs_rate_match() do not match the names used in the prototype.
This can be confusing to both humans and to code analysis tools, so
change the camelCase names used in the implementation to match the
prototype.

Change-Id: I87a85a7f1affa6bcad75f33f2bca6aa67475aa47
CRs-Fixed: 2365056
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>