qcacld-3.0: Add default NSS to rx_nss before flags calculation
The mcs flags that are sent to the upper layer are derived from mcs
table as present in the driver. Getting the correct values from the
table depends on providing the correct NSS, rate and rate flags.
Currently for the case of TX, all the parameters are passed correctly
but for RX, the NSS is not correct. The rx_nss is set to default value
of 0 where as it should be matched with the value as per the AP
capabilities.

Assign the NSS to the rx_nss as taken from vdev.

Change-Id: Idd579cd23c961d8b194b6b197bc65330be857746
CRs-Fixed: 2366271
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>