qcacld-3.0: Reset recovery in progress flag during probe
Currently in driver at load time if probe fails because
of any reason, as part of FW down, recovery in progress
flag is getting set which is getting reset only in re-init
path. Now when probe starts again as this was load time
failure, this flag is set and there are checks at many
places for this flag in driver where we just don't proceed
further if this flag is set.
One such example of such incident is reg_process_master_chan_list
api where regulatory event is not getting processed because this
flag is set.

To resolve this issue reset recovery in progress flag during probe.
There is no harm in resetting the flag during probe as probe start
indicates that FW is ready.

Change-Id: Ib681870f869060f4511563aa3255aeb0eb8f9f12
CRs-Fixed: 2365416
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>