qcacld-3.0: Check for the duplicate MAC during open adapter
Currently for STA mode in open adapter, locally administered
bit is getting reset, there is a possibility that after
resetting this locally administered bit this new MAC matches
to an existing MAC address which may result in multiple interfaces
to be up with same MAC address.

To avoid this issue, add a check for duplicate MAC address after
resetting locally administered bit.

Change-Id: I3ff8091df2c422bd89944ca39849c9961941e8c3
CRs-fixed: 2370349
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>