qcacmn: Rename sme_session_id in struct add_ts_param (2/2)
The sme_session_id field in struct add_ts_param is poorly named since
that is legacy MCL nomenclature, so there is a desire to rename it to
vdev_id to align with the converged nomenclature.

The first phase of renaming introduced a union so that both the old
and new name could be used by legacy projects, but only the new name
was used within the qcacmn project.

Now since all legacy projects have been updated to use the new name,
for the second phase remove the union, leaving behind just the new
name.

Change-Id: Ie86a3c0833899fbeacd458a53d9e23c6a67e7f12
CRs-Fixed: 2384157
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>