qcacld-3.0: Replace typedef tSirMacPowerCapInfo
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The tSirMacPowerCapInfo typedef
does not meet any of those criteria, so replace it (and the "tp"
variant) with a reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names
and so-called Hungarian notation, so in conjunction rename the
underlying struct to be in compliance.

Change-Id: I51fcb7d5d37d86b3ee9b5b104d48b0ad5354018b
CRs-Fixed: 2392403
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>