qcacld-3.0: Replace typedef tSirSmeStartBssReq
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The tSirSmeStartBssReq typedef does
not meet any of those criteria, so replace it (and the "tp" variant)
with a reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names and
so-called Hungarian notation, so in conjunction rename the underlying
struct to be in compliance.

Change-Id: I194d3de88ec3b877ca7e77abb42559c330b5ef69
CRs-Fixed: 2393732
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>