qcacld-3.0: Replace typedef tSirSmeTsmIEInd
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The tSirSmeTsmIEInd typedef does
not meet any of those criteria, so replace it (and the "tp" variant)
with a reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names
and so-called Hungarian notation, so in conjunction rename the
underlying struct to be in compliance.

Change-Id: I41b8b340fa3f781973d78edde3b54b7a5f0bd30d
CRs-Fixed: 2394254
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>