qcacld-3.0: Replace typedef tSirSmeJoinReq
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The tSirSmeJoinReq typedef does not
meet any of those criteria, so replace it (and the "tp" variant) with
a reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names
and so-called Hungarian notation, so in conjunction rename the
underlying struct to be in compliance.

Change-Id: Ie13a540223f8f08ad725b26cd55204f1221a4faa
CRs-Fixed: 2394996
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>