qcacld-3.0: Replace typedef tSirSmeJoinRsp
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The tSirSmeJoinRsp typedef does not
meet any of those criteria, so replace it (and the "tp" variant) with
a reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names
and so-called Hungarian notation, so in conjunction rename the
underlying struct to be in compliance.

Change-Id: I05b813c376cc837fa4580f5000a977ff0907c8a9
CRs-Fixed: 2394997
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>