qcacld-3.0: Replace typedef tSirSmeDisassocInd
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The tSirSmeDisassocInd typedef does
not meet any of those criteria, so replace it (and the "tp" variant)
with a reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names and
so-called Hungarian notation, so in conjunction rename the underlying
struct to be in compliance.

Change-Id: I97b38a01207d9060e1810a5747be2dd80d28de90
CRs-Fixed: 2395989
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>