qcacld-3.0: Replace session_id with vdev_id in hdd_adapter (3/5)
Originally struct hdd_adapter only used the legacy term session_id to
reference a session. To align with converged naming it is desirable to
change this to vdev_id.  In order to ease the transition change
Id92b1a7bb9721564ba01184809fec98be2143545 ("qcacld-3.0: Alias vdev_id
to session_id in hdd_adapter") replaced the session_id field with a
union that contains both a session_id and a vdev_id field. This allows
vdev_id and session_id to both be used to access the same data item.

To continue the transition replace the session_id field references
with vdev_in in the following files:
- wlan_hdd_ioctl.c
- wlan_hdd_main.c
- wlan_hdd_ota_test.c

Change-Id: I0411cf77bdc41fe72c9ebe85264a93957c66c191
CRs-Fixed: 2393154
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>