qcacld-3.0: Replace typedef tSirSmeMicFailureInd
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The tSirSmeMicFailureInd typedef
does not meet any of those criteria, so replace it (and the "tp"
variant) with a reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names and
so-called Hungarian notation, so in conjunction rename the underlying
struct to be in compliance.

Change-Id: Ic5cd76804a3e1462c1e92be3d1d2dd6911b108dc
CRs-Fixed: 2395996
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>