qcacld-3.0: Replace typedef tSirSmeMissedBeaconInd
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The tSirSmeMissedBeaconInd typedef
does not meet any of those criteria, so replace it (and the "tp"
variant) with a reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names and
so-called Hungarian notation, so in conjunction rename the underlying
struct to be in compliance.

Change-Id: I24f35dd3248c78c9f9cab3d724032ae1fa5e890d
CRs-Fixed: 2395997
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>