qcacld-3.0: Replace typedef roam_offload_synch_ind
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The roam_offload_synch_ind typedef
does not meet any of those criteria, so replace it (and the "tp"
variant) with a reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names and
so-called Hungarian notation, so in conjunction rename the underlying
struct to be in compliance.

Change-Id: Ieb91f75f4444873ab465d419409d46c05c65b091
CRs-Fixed: 2396068
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>