qcacld-3.0: Implement WAR for tdls_peer_update_state convergence
An upcoming set of changes will converge three sets of TDLS structs
into one. In order to allow the qcacld-3.0 changes to be made
independent of the qca-wifi-host-cmn changes change the call to
wmi_unified_update_tdls_peer_state_cmd() to use (void *) typecasting.

Note that this is a temporary change since the typecasting will be
completely removed when the convergence has been completed.

Change-Id: I59089575b2f5f7470ad60ca85247367c14e72bcd
CRs-Fixed: 2395336
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>