qcacmn: Change log level for qdf_mem_malloc failure
The non debug version of qdf_mem_malloc logs
allocation failures at WARN trace level
and the debug version uses ERROR trace level.

Also the default log level for QDF component is set to ERROR,
so with this, for failures in the non debug version
of qdf_mem_malloc, we would not see any logs.

Changing the log level to ERROR for failures logs in
the non debug version of qdf_mem_malloc.

Change-Id: I1e28cddfac8f909dec6e3bf6c3beb368953cd357
CRs-Fixed: 2397096
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>