qca-wifi: Remove Channels from ETSI PreCAC Required List when in NOL
When a channel is marked as radar detected and added to preCAC NOL
list, it is removed only from etsi precac done list, but not from etsi
precac required list. Since a channel that is part of NOL, cannot also be
a part of etsi precac required list, this needs to be corrected.

Therefore, once radar is detected on a channel, remove it from both etsi
precac required list and etsi precac done list.

Change-Id: I4d809a24acd4b73db254790e2698ffc5414cf0fe
CRs-Fixed: 2403489
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>