qcacld-3.0: Alias vdev_id to sessionId in csr_roam_session
Currently struct csr_roam_session uses the legacy term sessionId. To
align with converged naming it is desirable to change this to vdev_id,
but due to the large number of places where sessionId is used it will
be difficult to modify all occurences at once. In order to allow
changes to "trickle in" replace the sessionId field with a union that
contains both a sessionId and a vdev_id field. This will allow vdev_id
and sessionId to both be used to access the same data item.

Subsequent changes can then replace references to sessionId with
vdev_id, and then once all references have been converted, replace the
union with a single vdev_id field.

Change-Id: I67877b9a407bb07ec9f6516a765ee511c9c933e9
CRs-Fixed: 2404157
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>