qcacld-3.0: Update MC address list correctly
Currently while copying the MC address list, PMO
checks for the IPv6 router solicitation address. If this
address is found in the source list then this address is
ignored but the hole is created in the destination list.
As a result of this hole, FW does not filter the packets
for the last configured MC address in the list.

To mitigate this issue, add the correct logic of copying
the MC addresses from source list to destination list when
IPv6 router solicitation address is found in the source list.

Change-Id: If9412933d30259e7260e4d9e11cc514dec964355
CRs-Fixed: 2403419
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>