qcacld-3.0: Replace typedef tSmeConfigParams
The Linux Coding Style enumerates a few special cases where typedefs
are useful, but stresses "NEVER EVER use a typedef unless you can
clearly match one of those rules." The tSmeConfigParams typedef does
not meet any of those criteria, so replace it (and the "tp" variant)
with a reference to the underlying struct.

Further note the Linux Coding Style frowns upon mixed-case names and
so-called Hungarian notation, so in conjunction rename the underlying
struct to be in compliance.

Change-Id: I926cf67d87398782049bf9acbcd06f806a7cec29
CRs-Fixed: 2412715
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>