qcacld-3.0: Update set disable channel list command implementation
Currently driver disables channels received in
SET_DISABLE_CHANNEL_LIST command, only when sap is turned
on (on start bss). According to new requirement, channels
should be disabled as soon as the command is received.

To meet above expectation, update the set disable channel list
command functionality to disable the channels as soon as the
command is received.

CRs-fixed: 2424991
Change-Id: I596c21847e90efef9676f20df528aed77e927519
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>