qcacmn: Use cfreq to find if en302_502 is applicable
During radar table attach, the function to check if EN302_502
pattern is applicable requires actual center frequency of
the current operating channel as one of the inputs.
The caller of this function uses ieee80211 channel as the
argument in the function call.

Call function with actual center frequency instead of
ieee80211 channel.

Change-Id: I9e562863d73c53e4da046eede3e20d8ad9c8f98c
CRs-Fixed: 2454018
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>