qca-wifi: Free precac_entry from ETSI Precac required list on radar detect
When the CAC timer is in progress and radar is detected, a NULL pointer
dereference is observed when the radar detected channels are removed
from etsi required lists and added to NOL.

This issue is seen as the next etsi precac entry in the etsi required
list is freed instead of the current etsi precac entry.

This issue is fixed by freeing the current etsi precac entry in the
etsi precac required list.

Change-Id: I609174b23b0d6d2ea36ca26b011359dcf5876c3a
CRs-Fixed: 2463762
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>