qcacld-3.0: Filter the channel list according to acs cfg
Currently the driver checks whether the DFS channel is
in the range of acs channel list given, i.e the start
and end channel, which is not always correct as the channel
range does not imply that all channels in that range are present.

Fix is to explicitley check for the number of channels, and
compare each channel in the acs channel list with the DFS
channel.

Change-Id: Ib0b2d66dc325c340ab5cc837109a4f63e48ef4df
CRs-Fixed: 2466223
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>